Create PKGBUILD
This respects accepted practices in Arch and puts the release file where it should be.

Though it doesn't work as expected it seems to think that it is deno itself, when installed this way :

"error: unrecognized subcommand 'add'

Usage: deno [OPTIONS] [COMMAND]

For more information, try '--help'."

It doesn't work but it's a good start. After several hours and two tries, this is the best I can do.